Move specs for #filesystem_permissions
